data validation

+++ common/src/main/java/com/cqlybest/common/service/ProductService.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.cqlybest.common.bean.Product;
 import com.cqlybest.common.dao.ProductDao;
@@ -18,6 +19,7 @@ public class ProductService {
   /**
    * 编辑产品
    */
+  @Transactional
   public void edit(Product product) {
     product.setPublished(false);
     product.setLastUpdate(new Date());
@@ -27,8 +29,12 @@ public class ProductService {
   /**
    * 删除产品
    */
-  public void delete(Product product) {
-    productDao.delete(product);
+  @Transactional
+  public void delete(Integer id) {
+    Product product = productDao.findById(id);
+    if (product != null) {
+      productDao.delete(product);
+    }
   }
 
   public Long queryProductTotal() {
